Modified ExpenseDTO

+++ src/main/java/com/sokratis/ExpenseTracker/DTO/Category/CategoryDetailedDTO.java
@@ -3,14 +3,14 @@ package com.sokratis.ExpenseTracker.DTO.Category;
 import java.math.BigDecimal;
 import java.util.List;
 
-import com.sokratis.ExpenseTracker.DTO.ExpenseDTO;
+import com.sokratis.ExpenseTracker.DTO.Expense.ExpenseDTO;
 
 public record CategoryDetailedDTO(
-                Long id,
-                String name,
-                String color,
-                BigDecimal budgetLimit,
-                Long householdId,
-                Double spent,
-                List<ExpenseDTO> expenses) {
+        Long id,
+        String name,
+        String color,
+        BigDecimal budgetLimit,
+        Long householdId,
+        BigDecimal spent,
+        List<ExpenseDTO> expenses) {
 }
